Simplify PC handling and re-activate debugging code
For all instructions, the PC was manually advanced before, although for
most instructions (except for branching and popping PC instructions),
the PC simply advances by the length of the instruction at the PC. So we
simplify this by introducing a variable that indicates something special
happened to the PC during emulation (e.g., executed branch), but in all
other cases, the single stepping code automatically advances the PC.
This has the added advantage that non-implemented instructions are
auto-advanced as well (they're effectively a no-op now). During
implementation of this, I re-introduced new bugs (of course) which
caused me to re-activate the debugging code and properly implement it
with the new emu_ctx structure.