Implemented GDB stub to t2sim application, new-style tracing
Wrote a small test application that links against libthumb2sim and which
provides a gdb stub. Also polished the tracer application from scratch
that connects to this UNIX socket stub. This allows us to connect to
QEMU or thumb2sim, create traces using the exact identical process and
then compare them efficiently. The comparator application is also
included.  Note that with minimal modification this method can be
adapted to actually remote debug an actual harwdare target (but I'm
lacking one right now, so that's why I'm relying on QEMU).