Lots of refactoring and tracing
Further work on automatic creation of testcases. Found a few bugs in the
implementation by deliberately taking corner cases (e.g., adding
0xffffffff and 0x80000000, the V flag is mis-set). LOTS of changes to
the code. Added a Python-based instruction decoder so we need not only
rely on objdump, but can also decode the internals without invoking the
C code. Promising, but lots of work to be done.